do not remove too many chars on uncommenting
If the text is not consistently commented like this:

#- Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor
#  incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis
#  nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.
#  Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu
#  fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
#  culpa qui officia deserunt mollit anim id est laborum.

using gcu will remove the leading - on the first line. (`:set cms=#%s`)

That is, because in the first for loop, the l will be set to not include
the leading slash, however after the first iteration, it will be
overwritten and contain '# ' (e.g. 2 chars). Then on uncommenting,
the first two characters will be unconditionally removed.

Fix this, by remembering the shortest match to remove.

references tpope/vim-commentary#92